pyproject.toml
[project]
name = "gridarians"
version = "0.1.0"
description = "GRIDARIANS Project"

authors = [
    { name = "Can Gurkan", email = "cangurkan@me.com" }
]
dependencies = [
    "pydantic~=2.10.0",
    "groq~=0.12.0",
    "langchain~=0.3.7",
    "langchain-core~=0.3.19",
    "langchain-groq~=0.2.1",
    "pandas~=2.2.3",
    "seaborn~=0.13.2",
    "instructor~=1.6.4",
    "setuptools==59.5.0",
    "python-dotenv~=1.0.1",
    "anthropic~=0.45.0",
    "gin-config~=0.5.0",
    "langchain-anthropic~=0.3.7",
    "langchain-deepseek~=0.1.2",
    "httpx==0.27.2",
    "langgraph>=0.3.18",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.rye]
managed = true
dev-dependencies = []
